Change scroll behavior on lists
The scroll was not working with overflow hidden and ugly
with just overflow scroll. I use the new css property
scrollbar-width to not show up the scrollbar but to be able
to scroll anyway. I'm aware it is only supported by a few
browser versions but wanted to get used to a more natural
future way instead of this webkit-mess!